[not verified] publlicize: add message-box component
